ValidationTextFieldModel move into a separate file
